refactor(cli): fold pulse-sandbox into cli as src/sandbox
Slim-down phase 2. pulse-sandbox had exactly one consumer (cli's run_js
tool) and cli already carried a byte-identical copy of the fork runner
(src/sandbox-runner.ts) to satisfy the bundled runner lookup — the package
boundary was buying duplication, not reuse.

- move packages/pulse-sandbox/src -> packages/cli/src/sandbox (executor,
  tool, types, index, runner, tool.test)
- drop cli's duplicate src/sandbox-runner.ts; tsup runner entry now builds
  src/sandbox/runner.ts to the same dist/runner.cjs fork target
- executor resolveRunnerPath(): drop the require.resolve('pulse-sandbox')
  branch (dead once the package is gone); local dist candidates remain
- cli package.json: remove pulse-sandbox workspace dep (zod already direct);
  tsup: remove noExternal pulse-sandbox
- delete packages/pulse-sandbox; sync cli AGENTS/README, root READMEs,
  project-roadmap + requirement-to-mr skills

Verified: build green; test:core 471 tests green (sandbox specs moved into
cli: 8 files / 38 tests); check-harness harnessGaps: 0.
