+++ packages/cli/package.json
@@ -25,7 +25,6 @@
     "pulse-coder-engine": "workspace:*",
     "pulse-coder-memory-plugin": "workspace:*",
     "pulse-coder-orchestrator": "workspace:*",
-    "pulse-sandbox": "workspace:*",
     "react": "^19.2.5",
     "zod": "^4.3.6"
   },
