+++ packages/cli/src/sandbox/tool.test.ts
