--- packages/pulse-sandbox/harness/validate/validation.yaml
@@ -1,12 +1,0 @@
-version: 1
-
-pathRules:
-  - name: default
-    paths:
-      - src/**
-      - package.json
-      - tsconfig.json
-      - tsup.config.ts
-    required:
-      - pnpm --filter pulse-sandbox test
-      - pnpm --filter pulse-sandbox typecheck
