--- packages/pulse-sandbox/tsup.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: {
-    index: 'src/index.ts',
-    runner: 'src/runner.ts'
-  },
-  format: ['esm'],
-  dts: true,
-  clean: true,
-  splitting: false,
-  sourcemap: true,
-  target: 'es2022'
-});
